requirements.txt
flask==3.1.3
langchain==1.2.12
langchain-community==0.4.1
langchain-huggingface==1.2.1
sentence-transformers==5.2.3
faiss-cpu==1.13.2
transformers==4.41.2
torch==2.10.0
PyPDF2==3.0.1
python-docx==1.2.0
reportlab==4.4.10
gunicorn==25.1.0